+++ src/router/index.js
@@ -17,7 +17,7 @@ const router = createRouter({
     {
       path: '/rooms',
       name: 'rooms',
-      component: () => import('../History/Rooms_Table.vue')
+      component: () => import('../hostlify/Rooms_Table.vue')
     },
   ]
 })
